+++ src/components/calendar/index.ts
@@ -3,6 +3,7 @@ export { WeekView } from './WeekView';
 export { TwoWeekView } from './TwoWeekView';
 export { ThreeMonthView } from './ThreeMonthView';
 export { DayViewSideBySide } from './DayViewSideBySide';
+export { WeekVerticalView } from './WeekVerticalView';
 export { CalendarFilterPopover } from './CalendarFilterPopover';
 
 export type { MonthViewProps } from './MonthView';
@@ -10,3 +11,4 @@ export type { WeekViewProps } from './WeekView';
 export type { TwoWeekViewProps } from './TwoWeekView';
 export type { ThreeMonthViewProps } from './ThreeMonthView';
 export type { DayViewSideBySideProps } from './DayViewSideBySide';
+export type { WeekVerticalViewProps } from './WeekVerticalView';